backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.6.4
pydantic-settings==2.2.1
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# AI / NLP
google-generativeai==0.5.2
sentence-transformers==2.7.0
chromadb==0.5.0
torch==2.2.2
transformers==4.39.3

# OCR
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0

# HTTP
httpx==0.27.0
requests==2.31.0

# Utilities
python-jose==3.3.0
passlib[bcrypt]==1.7.4
